requirements.txt
PyMuPDF==1.23.8
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3